# scan-engine/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "scan-to-map-studio API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects": {
      "get": {
        "summary": "Api List Projects",
        "operationId": "api_list_projects_api_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectEntry"
                  },
                  "type": "array",
                  "title": "Response Api List Projects Api Projects Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Api Create Project",
        "operationId": "api_create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_create_project_api_projects_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{name}": {
      "delete": {
        "summary": "Api Delete Project",
        "operationId": "api_delete_project_api_projects__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_api_create_project_api_projects_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Body_api_create_project_api_projects_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreated": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectCreated"
      },
      "ProjectEntry": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "steps": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Steps",
            "default": {}
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "mtime": {
            "type": "number",
            "title": "Mtime"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "name",
          "phase",
          "mtime"
        ],
        "title": "ProjectEntry",
        "description": "GET /api/projects item (== studio.project.list_projects)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}